Update helper.py
Added changes  related to the mismatch between device memory locations. In other words, your model and data are not in the same location (one is in CPU and the other is in GPU, or vice versa). This often occurs when we forget to move both the model and the input data to the same device (either both on CPU or both on GPU).